% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_model}
\alias{get_default_model}
\title{获取默认模型名称}
\usage{
get_default_model(provider = "deepseek", custom_model = NULL)
}
\arguments{
\item{provider}{API 提供商类型}

\item{custom_model}{自定义模型名称}
}
\value{
模型名称字符串
}
\description{
获取默认模型名称
}
\keyword{internal}
