% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalescing}
\alias{null_coalescing}
\alias{\%||\%}
\title{默认值替换符 (Null coalescing operator)}
\usage{
x \%||\% y
}
\arguments{
\item{x}{左侧变量}

\item{y}{右侧默认值}
}
\value{
如果 x 为 NULL 则返回 y，否则返回 x
}
\description{
默认值替换符 (Null coalescing operator)
}
