% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_with_deepseek}
\alias{translate_with_deepseek}
\title{使用 DeepSeek API 进行摘要翻译}
\usage{
translate_with_deepseek(text, api_key, max_tokens = 10000)
}
\arguments{
\item{text}{需要翻译的英文文本}

\item{api_key}{DeepSeek API 密钥}

\item{max_tokens}{最大 Token 数}
}
\description{
使用 DeepSeek API 进行摘要翻译
}
